docs: add jekyll configuration file
This commit introduces a new Jekyll configuration file, jekyll.yml, which is essential for customizing the build process of the Jekyll site. The file includes settings for the base URL, title, and other metadata that will be used to generate the site.

The configuration is set up with a base URL of 'https://example.com' and a site title of 'My Jekyll Blog'. Additional options may be added in the future to further customize the site's appearance and functionality.

Signed-off-by: BG4JTS <157703235+BG4JTS@users.noreply.github.com>